requirements.txt
Django==4.0.5
djangorestframework==3.14.0
djangorestframework_simplejwt==4.8.0
drf_spectacular==0.26.2
python-dotenv==1.0.0
rest_framework_simplejwt==0.0.2
